model: return value from switch_to_master

+++ model.h
@@ -116,7 +116,7 @@ public:
 	unsigned int get_num_threads() const;
 	Thread * get_current_thread() const;
 
-	int switch_to_master(ModelAction *act);
+	uint64_t switch_to_master(ModelAction *act);
 	ClockVector * get_cv(thread_id_t tid) const;
 	ModelAction * get_parent_action(thread_id_t tid) const;
 	void check_promises_thread_disabled();
